refactor: using new events data model (wip)
