+++ src/App.tsx
@@ -1,11 +1,12 @@
 import './App.css'
 
+import Game from 'pages/Game/Game'
+import Home from 'pages/Home/Home'
 import Run from 'pages/Run/Run'
+import Team from 'pages/Team/Team'
 import { BrowserRouter, Route, Routes } from 'react-router-dom'
 
 import getGameData from './data/index'
-import Game from './pages/Game/Game'
-import Home from './pages/Home/Home'
 
 function App() {
   const gameData = getGameData()
@@ -16,8 +17,7 @@ function App() {
         <Route path="/" element={<Home data={gameData} />} />
         <Route path="/:gameKey" element={<Game data={gameData} />} />
         <Route path="/:gameKey/:runKey" element={<Run data={gameData} />} />
-        {/* <Route path="/:gameKey/:runKey/:teamKey" element={<Team data={gameData} />} />
-         */}
+        <Route path="/:gameKey/:runKey/:teamKey" element={<Team data={gameData} />} />
       </Routes>
     </BrowserRouter>
   )
