+++ src/data/test-game.ts
@@ -735,7 +735,13 @@ export default {
           type: EventType.ChallengeCompleted,
           timestamp: relativeDate(-30, 14, 0),
           team: 'team-a',
-          challenge: 'Take pic with Van Gogh',
+          challenge: 'Spot a canal',
+        },
+        {
+          type: EventType.ChallengeCompleted,
+          timestamp: relativeDate(-30, 15, 0),
+          team: 'team-a',
+          challenge: 'Spot a cheese shop',
         },
         { type: EventType.Finish, timestamp: relativeDate(-30, 18, 0) },
       ],
